+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: jalevesq <jalevesq@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 14:31:24 by jalevesq          #+#    #+#             */
-/*   Updated: 2023/04/05 14:40:06 by jalevesq         ###   ########.fr       */
+/*   Updated: 2023/04/05 17:31:43 by jalevesq         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -96,24 +96,35 @@ void				ft_ctrlc(int sig);
 void				ft_title(void);
 
 int					ft_is_cmd(t_token *token);
-// Every function for Executor
+
+/* *** EVERY FUNCTION FOR EXECUTOR *** */
+
+/* MAIN FUNCTION */
 void				ft_executor(t_token *token, char **envp);
 void				ft_process_child(t_child *c, t_token *tmp, int *pid);
 
+/* CHILD ERROR FUNCTION */
+void				ft_child_error(t_token *token, t_child *c, int flag);
+
+/* REDIRECTION && HERE_DOC */
+void				ft_great_child(t_child *child, t_token *token, int great);
+void				ft_less_child(t_child *child, t_token *token, int less);
+void				ft_pipe_child(t_child *child, t_token *token);
+
+void				ft_heredoc(t_token *token, t_child *child);
+int					ft_heredoc_nbr(t_token *t);
+int					ft_is_doc_last(t_token *token);
+
+/* UTILS FOR EXECUTOR */
 int					*ft_set_pipe(t_child *child);
 int					cmd_counter(t_token *token);
 void				ft_wait(pid_t *pid, int cmd_nbr);
 void				ft_close_fd(int *fd_array, int cmd_nbr);
 int					ft_mark_count(t_token *token, int type);
-
-void				ft_child_error(t_token *token, t_child *c, int flag);
-
-void				ft_great_child(t_child *child, t_token *token, int great);
-void				ft_less_child(t_child *child, t_token *token, int less);
-void				ft_pipe_child(t_child *child, t_token *token);
-
 char				*find_cmd_path(char **cmd, char **path);
 char				**find_path(char **envp);
+
+/*********************************************************/
 
 // Free & End function.
 void				*ft_free_double(char **str);
@@ -121,8 +132,4 @@ void				ft_end_list(t_token *token);
 void				ft_free_list(t_token *token);
 void				ft_error(int flag);
 void				ft_free_exec(char **cmd, char *cmd_path);
-// void				free_container(t_cmd *container);
-// void				error_cmd_path(t_cmd *container);
-// void				free_cmd(t_cmd *container);
-
 #endif
